Don't stream wals with pg_basebackup
and start pg_receivewal before starting pg_basebackup

+++ postgres-appliance/basebackup.sh
@@ -19,45 +19,46 @@ done
 
 [[ -z $DATA_DIR || -z $CONNSTR || ! $RETRIES =~ ^[1-9]$ ]] && exit 1
 
-function receivewal() {
-    if which pg_receivewal &> /dev/null; then
-        PG_RECEIVEWAL=pg_receivewal
-    else
-        PG_RECEIVEWAL=pg_receivexlog
-    fi
+if which pg_receivewal &> /dev/null; then
+    PG_RECEIVEWAL=pg_receivewal
+    PG_BASEBACKUP_OPTS="-X none"
+else
+    PG_RECEIVEWAL=pg_receivexlog
+    PG_BASEBACKUP_OPTS=""
+fi
+
+readonly WAL_FAST=$(dirname $DATA_DIR)/wal_fast
+mkdir -p $WAL_FAST
+
+# make sure that there is no receivewal running
+exec 9>$WAL_FAST/receivewal.lock
+if flock -x -n 9; then
     $PG_RECEIVEWAL --directory="${WAL_FAST}" --dbname="${CONNSTR}" &
     receivewal_pid=$!
 
     # run pg_receivewal until postgres will not start streaming
-    while ! ps ax | grep -qE '[w]al receiver process\s+streaming'; do
-        # exit if pg_receivewal is not running
-        kill -0 $receivewal_pid && sleep 1 || exit
-    done
+    (
+        while ! ps ax | grep -qE '[w]al receiver process\s+streaming'; do
+            # exit if pg_receivewal is not running
+            kill -0 $receivewal_pid && sleep 1 || exit
+        done
 
-    kill $receivewal_pid && sleep 1
-    rm -f ${WAL_FAST}/*
-}
+        kill $receivewal_pid && sleep 1
+        rm -f ${WAL_FAST}/*
+    )&
+fi
 
 ATTEMPT=0
 while [[ $((ATTEMPT++)) -le $RETRIES ]]; do
     rm -fr "${DATA_DIR}"
-    pg_basebackup --pgdata="${DATA_DIR}" -X stream --dbname="${CONNSTR}"
+    pg_basebackup --pgdata="${DATA_DIR}" ${PG_BASEBACKUP_OPTS} --dbname="${CONNSTR}"
     EXITCODE=$?
     if [[ $EXITCODE == 0 ]]; then
-        WAL_FAST=$(dirname $DATA_DIR)/wal_fast
-
-        WAL_DIR=${DATA_DIR}/pg_wal
-        [[ -d ${WAL_DIR} ]] || WAL_DIR=${DATA_DIR}/pg_xlog
-
-        rm -fr $WAL_FAST $WAL_DIR/archive_status
-
-        mv $WAL_DIR $WAL_FAST
-        mkdir $WAL_DIR
-
-        receivewal &
         break
     elif [[ $ATTEMPT -le $RETRIES ]]; then
         sleep $((ATTEMPT*10))
     fi
 done
+
+[[ $EXITCODE != 0 && ! -z $receivewal_pid ]] && kill $receivewal_pid
 exit $EXITCODE
